pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slide-generator"
version = "0.1.0"
description = "AI-powered slide deck generation tool using natural language"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics :: Presentation",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"
dependencies = [
    # Core slide generator dependencies
    "databricks-sdk>=0.20.0",
    "pathlib",
    "typing-extensions",
    "pydantic>=2.0.0",
    "beautifulsoup4>=4.12.0",
    
    # FastAPI backend dependencies
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0", 
    "python-multipart>=0.0.6",
    "websockets>=11.0",
    
    # Unity Catalog and Databricks dependencies
    "unitycatalog-ai",
    "databricks-connect",
    "openai",
    "databricks-sdk[openai]",
    
    # LangChain dependencies
    "langchain-core>=0.1.0",
    "langchain-databricks>=0.1.0", 
    "langchain>=0.1.0",
    "langgraph>=0.1.0",
    
    # Additional data handling
    "pandas>=2.0.0",
]

[project.optional-dependencies]
# Legacy gradio support (deprecated)
gradio-legacy = [
    "gradio>=4.0.0",
]
dev = [
    # Existing dev dependencies
    "pytest>=7.0",
    "pytest-cov",
    "black>=22.0", 
    "flake8>=5.0",
    "mypy>=1.0",
    "pre-commit>=2.20",
    "isort>=5.10",
    
    # Additional testing dependencies for Phase 1.5
    "pytest-asyncio>=0.21.0",  # For FastAPI async tests
    "pytest-mock>=3.10.0",     # For mocking Databricks responses  
    "httpx>=0.24.0",           # For FastAPI TestClient
    "pytest-html>=3.1.0",     # For HTML test reports
    "aiohttp>=3.8.0",         # For load testing
    "psutil>=5.9.0",          # For memory monitoring tests
]
docs = [
    "mkdocs>=1.4",
    "mkdocs-material>=8.5",
    "mkdocstrings[python]>=0.19",
]

[project.urls]
Homepage = "https://github.com/your-username/slide-generator"
Documentation = "https://slide-generator.readthedocs.io/"
Repository = "https://github.com/your-username/slide-generator.git"
Issues = "https://github.com/your-username/slide-generator/issues"

[project.scripts]
slide-generator = "slide_generator.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
slide_generator = ["*.md", "*.txt", "*.yml", "*.yaml"]

# Black configuration
[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# MyPy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    # "gradio.*",  # REMOVED: Gradio frontend deprecated
    "databricks.*",
    "langchain.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=slide_generator",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
    "performance: marks tests as performance/load tests",
    "databricks: marks tests that require Databricks authentication",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/slide_generator"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

